+++ runtime/nms/1.19.2/src/main/java/net/cocoonmc/runtime/v1_19_R1/TransformFactory.java
@@ -100,7 +100,7 @@ public class TransformFactory {
         }
     }
 
-    public static Vec3 convertToVanilla(net.cocoonmc.core.math.Vector3f vector) {
+    public static Vec3 convertToVanilla(net.cocoonmc.core.math.Vector3d vector) {
         return new Vec3(vector.getX(), vector.getY(), vector.getZ());
     }
